pyproject.toml
[tool.poetry]
name = "pilgrim"
version = "0.2.8"
description = ""
authors = ["Byeonghoon Yoo <bh322yoo@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.63.0"
Jinja2 = "^2.11.3"
aiofiles = "^0.6.0"
aiocache = "^0.11.1"
msgpack = "^1.0.2"
ujson = "^4.0.2"
tortoise-orm = "^0.16.21"
ciso8601 = "^2.1.3"
python-rapidjson = "^1.0"
ddtrace = "^0.47.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2"
black = "^20.8b1"
uvicorn = "^0.13.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target-version = ['py38']
skip-string-normalization = true
